// src/pages/bookv2/hooks/useBookingFormData.generated.ts
export type BookingContactInfoValuesFragment = { email: string, firstName: string, id: number, lastName: string, phone: string };

export type BookingContactInfoQueryVariables = Types.Exact<{
  bookingUid?: Types.InputMaybe<Types.Scalars['String']>;
}>;


export type BookingContactInfoQuery = { bookingContactInfo?: { email: string, firstName: string, id: number, lastName: string, phone: string } | null };

export type BookingParticipantDetailsValuesFragment = { bookingFareTypeId: number, bookingFareTypeName: string, firstName?: string | null, id: number, lastName?: string | null };